requirements.txt
absl-py==0.8.1
alembic==1.2.1
aniso8601==7.0.0
astor==0.8.0
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.2.0
backcall==0.1.0
bleach==3.1.0
Click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-GraphQL==2.0.0
gast==0.2.2
gitdb2==2.0.6
google-pasta==0.1.7
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
grpcio==1.24.1
h5py==2.10.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
loguru==0.3.2
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pkg-resources==0.0.0
pluggy==0.13.0
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.0
pyaml==19.4.1
Pygments==2.4.2
pylint==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.0.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
rope==0.14.0
Rx==1.6.1
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smmap2==2.0.5
SQLAlchemy==1.3.10
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
treelib==1.5.5
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
